Updated Class 2 File
# Class Announcement

Hi everyone, thank you for the last session.

I have been a bit ill this week and am now catching up on the follow-ups.

## Next Session

**Date & Time:** 10.12 from 10:00 til 13:00  
**Location:** Gamelab (324)

## Guest Lecture

**Speaker:** Mario Guzman  
**Topic:** LLMs and chatbots  
**Date & Time:** 11.12 at 11:00  
**Location:** R103

## Todos till next session

Integrate an animation into the composition you have already created. You can do this by taking the class file as an example and doing one of the following:

1. **Use an LFO** to animate an aspect of your composition, such as transform on the X axis, scale it on the X and Y axes, or on each axis, etc.

2. **Use keyframe animation** to animate a component in your sketch, such as translate, scale, levels, etc.

3. **Use a noise operator** to animate colour or any other parameter you think you want to.

4. **Be as deliberate as you can** — try experimenting with creating a 3 or 2-step composition and an idea of a concept, e.g.
   - **Concept:** I want to experiment with colour theory and repetitive systems.
   - **Composition:**
     1. Scale a rectangle within a defined interval.
     2. Change the rotation of the rectangle within a defined interval to create a sense of proportion.
     3. Change colour and overlay different colours to study the effects of different colour compositions.

Check out the example file. On Composition and Concept:

**What would be one element you would like to have a state change in?** For example: from still composition to animated composition, from a colour to b colour, from a scale to b scale, etc.

We will create a trigger in the next session based on the system to switch states, and we will look into [MediaPipe](https://github.com/torinmb/mediapipe-touchdesigner) to see how to use it for such a trigger.